ALL MEASUREMENT UNITS SWITCHED TO INTEGERS: Switch from using float/double/m/km/sec units to int/long/mm/msec units throughout. New file format is not backward compatible with the old format.

+++ trunk/src/edu/gatech/lbs/core/world/roadnet/RoadJunctionDistance.java
@@ -1,16 +1,20 @@
+// Copyright (c) 2012, Georgia Tech Research Corporation
+// Authors:
+//   Peter Pesti (pesti@gatech.edu)
+//
 package edu.gatech.lbs.core.world.roadnet;
 
 public class RoadJunctionDistance implements Comparable<RoadJunctionDistance> {
   public RoadJunction junction;
-  public double distance;
+  public int distance;
 
-  public RoadJunctionDistance(RoadJunction junction, double distance) {
+  public RoadJunctionDistance(RoadJunction junction, int distance) {
     this.junction = junction;
     this.distance = distance;
   }
 
   public int compareTo(RoadJunctionDistance d2) {
-    double diff = distance - d2.distance;
+    int diff = distance - d2.distance;
     // primary method of comparison is distance, but use junctionId as tiebreaker:
     if (diff == 0) {
       return junction.getId() - d2.junction.getId();
