+++ trunk/src/edu/gatech/lbs/core/world/roadnet/route/IRoute.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2009, Georgia Tech Research Corporation
+// Copyright (c) 2012, Georgia Tech Research Corporation
 // Authors:
 //   Peter Pesti (pesti@gatech.edu)
 //
@@ -7,7 +7,7 @@ package edu.gatech.lbs.core.world.roadnet.route;
 import edu.gatech.lbs.core.vector.RoadnetVector;
 
 public interface IRoute {
-  public double getLength();
+  public long getLength();
 
   public RoadnetVector getSource();
